chore(storage): split storage reader worker
