+++ src/EventStore.Core/Services/Storage/StorageReaderWorker.cs
@@ -26,7 +26,7 @@ public abstract class StorageReaderWorker
 	protected static readonly ILogger Log = Serilog.Log.ForContext<StorageReaderWorker>();
 }
 
-public class StorageReaderWorker<TStreamId> :
+public partial class StorageReaderWorker<TStreamId> :
 	StorageReaderWorker,
 	IAsyncHandle<ClientMessage.ReadEvent>,
 	IAsyncHandle<ClientMessage.ReadStreamEventsBackward>,
@@ -75,523 +75,6 @@ public class StorageReaderWorker<TStreamId> :
 		_virtualStreamReader = virtualStreamReader;
 	}
 
-	async ValueTask IAsyncHandle<ClientMessage.ReadEvent>.HandleAsync(ClientMessage.ReadEvent msg,
-		CancellationToken token)
-	{
-		if (msg.CancellationToken.IsCancellationRequested)
-		{
-			return;
-		}
-
-		if (msg.Expires < DateTime.UtcNow)
-		{
-			if (LogExpiredMessage(msg.Expires))
-			{
-				Log.Debug(
-					"Read Event operation has expired for Stream: {stream}, Event Number: {eventNumber}. Operation Expired at {expiryDateTime}",
-					msg.EventStreamId, msg.EventNumber, msg.Expires);
-			}
-
-			return;
-		}
-
-		using var cts = Multiplex(ref token, msg);
-		try
-		{
-			var ev = await ReadEvent(msg, token);
-			msg.Envelope.ReplyWith(ev);
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token && cts.IsTimedOut)
-		{
-			if (LogExpiredMessage(msg.Expires))
-			{
-				Log.Debug(
-					"Read Event operation has expired for Stream: {stream}, Event Number: {eventNumber}. Operation Expired at {expiryDateTime}",
-					msg.EventStreamId, msg.EventNumber, msg.Expires);
-			}
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token)
-		{
-			throw new OperationCanceledException(null, ex, cts.CancellationOrigin);
-		}
-	}
-
-	async ValueTask IAsyncHandle<ClientMessage.ReadStreamEventsForward>.HandleAsync(
-		ClientMessage.ReadStreamEventsForward msg, CancellationToken token)
-	{
-		if (msg.CancellationToken.IsCancellationRequested)
-		{
-			return;
-		}
-
-		if (msg.Expires < DateTime.UtcNow)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.ReadStreamEventsForwardCompleted(
-					msg.CorrelationId, msg.EventStreamId, msg.FromEventNumber, msg.MaxCount,
-					ReadStreamResult.Expired,
-					ResolvedEvent.EmptyArray, default, default, default, default, default, default, default));
-			}
-
-			if (LogExpiredMessage(msg.Expires))
-			{
-				Log.Debug(
-					"Read Stream Events Forward operation has expired for Stream: {stream}, From Event Number: {fromEventNumber}, Max Count: {maxCount}. Operation Expired at {expiryDateTime}",
-					msg.EventStreamId, msg.FromEventNumber, msg.MaxCount, msg.Expires);
-			}
-
-			return;
-		}
-
-		ClientMessage.ReadStreamEventsForwardCompleted res;
-		using var cts = Multiplex(ref token, msg);
-		try
-		{
-			res = _virtualStreamReader.CanReadStream(msg.EventStreamId)
-				? await _virtualStreamReader.ReadForwards(msg, token)
-				: await ReadStreamEventsForward(msg, token);
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token && cts.IsTimedOut)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.ReadStreamEventsForwardCompleted(
-					msg.CorrelationId, msg.EventStreamId, msg.FromEventNumber, msg.MaxCount,
-					ReadStreamResult.Expired,
-					ResolvedEvent.EmptyArray, default, default, default, default, default, default, default));
-			}
-
-			if (LogExpiredMessage(msg.Expires))
-			{
-				Log.Debug(
-					"Read Stream Events Forward operation has expired for Stream: {stream}, From Event Number: {fromEventNumber}, Max Count: {maxCount}. Operation Expired at {expiryDateTime}",
-					msg.EventStreamId, msg.FromEventNumber, msg.MaxCount, msg.Expires);
-			}
-
-			return;
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token)
-		{
-			throw new OperationCanceledException(null, ex, cts.CancellationOrigin);
-		}
-
-		switch (res.Result)
-		{
-			case ReadStreamResult.Success:
-			case ReadStreamResult.NoStream:
-			case ReadStreamResult.NotModified:
-				if (msg.LongPollTimeout.HasValue && res.FromEventNumber > res.LastEventNumber)
-				{
-					_publisher.Publish(new SubscriptionMessage.PollStream(
-						msg.EventStreamId, res.TfLastCommitPosition, res.LastEventNumber,
-						DateTime.UtcNow + msg.LongPollTimeout.Value, msg));
-				}
-				else
-				{
-					msg.Envelope.ReplyWith(res);
-				}
-
-				break;
-			case ReadStreamResult.StreamDeleted:
-			case ReadStreamResult.Error:
-			case ReadStreamResult.AccessDenied:
-				msg.Envelope.ReplyWith(res);
-				break;
-			default:
-				throw new ArgumentOutOfRangeException(
-					$"Unknown ReadStreamResult: {res.Result}");
-		}
-	}
-
-	async ValueTask IAsyncHandle<ClientMessage.ReadStreamEventsBackward>.HandleAsync(
-		ClientMessage.ReadStreamEventsBackward msg, CancellationToken token)
-	{
-		if (msg.CancellationToken.IsCancellationRequested)
-		{
-			return;
-		}
-
-		if (msg.Expires < DateTime.UtcNow)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.ReadStreamEventsBackwardCompleted(
-					msg.CorrelationId, msg.EventStreamId, msg.FromEventNumber, msg.MaxCount,
-					ReadStreamResult.Expired,
-					ResolvedEvent.EmptyArray, default, default, default, default, default, default, default));
-			}
-
-			if (LogExpiredMessage(msg.Expires))
-			{
-				Log.Debug(
-					"Read Stream Events Backward operation has expired for Stream: {stream}, From Event Number: {fromEventNumber}, Max Count: {maxCount}. Operation Expired at {expiryDateTime}",
-					msg.EventStreamId, msg.FromEventNumber, msg.MaxCount, msg.Expires);
-			}
-
-			return;
-		}
-
-		using var cts = Multiplex(ref token, msg);
-		try
-		{
-			var res = _virtualStreamReader.CanReadStream(msg.EventStreamId)
-				? await _virtualStreamReader.ReadBackwards(msg, token)
-				: await ReadStreamEventsBackward(msg, token);
-
-			msg.Envelope.ReplyWith(res);
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token && cts.IsTimedOut)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.ReadStreamEventsBackwardCompleted(
-					msg.CorrelationId, msg.EventStreamId, msg.FromEventNumber, msg.MaxCount,
-					ReadStreamResult.Expired,
-					ResolvedEvent.EmptyArray, default, default, default, default, default, default, default));
-			}
-
-			if (LogExpiredMessage(msg.Expires))
-			{
-				Log.Debug(
-					"Read Stream Events Backward operation has expired for Stream: {stream}, From Event Number: {fromEventNumber}, Max Count: {maxCount}. Operation Expired at {expiryDateTime}",
-					msg.EventStreamId, msg.FromEventNumber, msg.MaxCount, msg.Expires);
-			}
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token)
-		{
-			throw new OperationCanceledException(null, ex, cts.CancellationOrigin);
-		}
-	}
-
-	async ValueTask IAsyncHandle<ClientMessage.ReadAllEventsForward>.HandleAsync(ClientMessage.ReadAllEventsForward msg,
-		CancellationToken token)
-	{
-		if (msg.CancellationToken.IsCancellationRequested)
-		{
-			return;
-		}
-
-		if (msg.Expires < DateTime.UtcNow)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.ReadAllEventsForwardCompleted(
-					msg.CorrelationId, ReadAllResult.Expired,
-					default, ResolvedEvent.EmptyArray, default, default, default,
-					currentPos: new TFPos(msg.CommitPosition, msg.PreparePosition),
-					TFPos.Invalid, TFPos.Invalid, default));
-			}
-
-			if (LogExpiredMessage(msg.Expires))
-			{
-				Log.Debug(
-					"Read All Stream Events Forward operation has expired for C:{commitPosition}/P:{preparePosition}. Operation Expired at {expiryDateTime}",
-					msg.CommitPosition, msg.PreparePosition, msg.Expires);
-			}
-
-			return;
-		}
-
-		using var cts = Multiplex(ref token, msg);
-		try
-		{
-			var res = await ReadAllEventsForward(msg, token);
-			switch (res.Result)
-			{
-				case ReadAllResult.Success:
-					if (msg.LongPollTimeout.HasValue && res.IsEndOfStream && res.Events.Count is 0)
-					{
-						_publisher.Publish(new SubscriptionMessage.PollStream(
-							SubscriptionsService.AllStreamsSubscriptionId, res.TfLastCommitPosition, null,
-							DateTime.UtcNow + msg.LongPollTimeout.Value, msg));
-					}
-					else
-					{
-						msg.Envelope.ReplyWith(res);
-					}
-
-					break;
-				case ReadAllResult.NotModified:
-					if (msg.LongPollTimeout.HasValue && res.IsEndOfStream &&
-						res.CurrentPos.CommitPosition > res.TfLastCommitPosition)
-					{
-						_publisher.Publish(new SubscriptionMessage.PollStream(
-							SubscriptionsService.AllStreamsSubscriptionId, res.TfLastCommitPosition, null,
-							DateTime.UtcNow + msg.LongPollTimeout.Value, msg));
-					}
-					else
-					{
-						msg.Envelope.ReplyWith(res);
-					}
-
-					break;
-				case ReadAllResult.Error:
-				case ReadAllResult.AccessDenied:
-				case ReadAllResult.InvalidPosition:
-					msg.Envelope.ReplyWith(res);
-					break;
-				default:
-					throw new ArgumentOutOfRangeException($"Unknown ReadAllResult: {res.Result}");
-			}
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token && cts.IsTimedOut)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.ReadAllEventsForwardCompleted(
-					msg.CorrelationId, ReadAllResult.Expired,
-					default, ResolvedEvent.EmptyArray, default, default, default,
-					currentPos: new TFPos(msg.CommitPosition, msg.PreparePosition),
-					TFPos.Invalid, TFPos.Invalid, default));
-			}
-
-			if (LogExpiredMessage(msg.Expires))
-			{
-				Log.Debug(
-					"Read All Stream Events Forward operation has expired for C:{commitPosition}/P:{preparePosition}. Operation Expired at {expiryDateTime}",
-					msg.CommitPosition, msg.PreparePosition, msg.Expires);
-			}
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token)
-		{
-			throw new OperationCanceledException(null, ex, cts.CancellationOrigin);
-		}
-	}
-
-	async ValueTask IAsyncHandle<ClientMessage.ReadAllEventsBackward>.HandleAsync(
-		ClientMessage.ReadAllEventsBackward msg, CancellationToken token)
-	{
-		if (msg.CancellationToken.IsCancellationRequested)
-		{
-			return;
-		}
-
-		if (msg.Expires < DateTime.UtcNow)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.ReadAllEventsBackwardCompleted(
-					msg.CorrelationId, ReadAllResult.Expired,
-					default, ResolvedEvent.EmptyArray, default, default, default,
-					currentPos: new TFPos(msg.CommitPosition, msg.PreparePosition),
-					TFPos.Invalid, TFPos.Invalid, default));
-			}
-
-			if (LogExpiredMessage(msg.Expires))
-			{
-				Log.Debug(
-					"Read All Stream Events Backward operation has expired for C:{commitPosition}/P:{preparePosition}. Operation Expired at {expiryDateTime}",
-					msg.CommitPosition, msg.PreparePosition, msg.Expires);
-			}
-
-			return;
-		}
-
-		using var cts = Multiplex(ref token, msg);
-		try
-		{
-			msg.Envelope.ReplyWith(await ReadAllEventsBackward(msg, token));
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token && cts.IsTimedOut)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.ReadAllEventsBackwardCompleted(
-					msg.CorrelationId, ReadAllResult.Expired,
-					default, ResolvedEvent.EmptyArray, default, default, default,
-					currentPos: new TFPos(msg.CommitPosition, msg.PreparePosition),
-					TFPos.Invalid, TFPos.Invalid, default));
-			}
-
-			if (LogExpiredMessage(msg.Expires))
-			{
-				Log.Debug(
-					"Read All Stream Events Backward operation has expired for C:{commitPosition}/P:{preparePosition}. Operation Expired at {expiryDateTime}",
-					msg.CommitPosition, msg.PreparePosition, msg.Expires);
-			}
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token)
-		{
-			throw new OperationCanceledException(null, ex, cts.CancellationOrigin);
-		}
-	}
-
-	async ValueTask IAsyncHandle<ClientMessage.FilteredReadAllEventsForward>.HandleAsync(
-		ClientMessage.FilteredReadAllEventsForward msg, CancellationToken token)
-	{
-		if (msg.CancellationToken.IsCancellationRequested)
-		{
-			return;
-		}
-
-		if (msg.Expires < DateTime.UtcNow)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.FilteredReadAllEventsForwardCompleted(
-					msg.CorrelationId, FilteredReadAllResult.Expired,
-					default, ResolvedEvent.EmptyArray, default, default, default,
-					currentPos: new TFPos(msg.CommitPosition, msg.PreparePosition),
-					TFPos.Invalid, TFPos.Invalid, default, default, default));
-			}
-
-			Log.Debug(
-				"Read All Stream Events Forward Filtered operation has expired for C:{0}/P:{1}. Operation Expired at {2}",
-				msg.CommitPosition, msg.PreparePosition, msg.Expires);
-			return;
-		}
-
-		using var cts = Multiplex(ref token, msg);
-		try
-		{
-			var res = await FilteredReadAllEventsForward(msg, token);
-			switch (res.Result)
-			{
-				case FilteredReadAllResult.Success:
-					if (msg.LongPollTimeout.HasValue && res.IsEndOfStream && res.Events.Count is 0)
-					{
-						_publisher.Publish(new SubscriptionMessage.PollStream(
-							SubscriptionsService.AllStreamsSubscriptionId, res.TfLastCommitPosition, null,
-							DateTime.UtcNow + msg.LongPollTimeout.Value, msg));
-					}
-					else
-					{
-						msg.Envelope.ReplyWith(res);
-					}
-
-					break;
-				case FilteredReadAllResult.NotModified:
-					if (msg.LongPollTimeout.HasValue && res.IsEndOfStream &&
-						res.CurrentPos.CommitPosition > res.TfLastCommitPosition)
-					{
-						_publisher.Publish(new SubscriptionMessage.PollStream(
-							SubscriptionsService.AllStreamsSubscriptionId, res.TfLastCommitPosition, null,
-							DateTime.UtcNow + msg.LongPollTimeout.Value, msg));
-					}
-					else
-					{
-						msg.Envelope.ReplyWith(res);
-					}
-
-					break;
-				case FilteredReadAllResult.Error:
-				case FilteredReadAllResult.AccessDenied:
-				case FilteredReadAllResult.InvalidPosition:
-					msg.Envelope.ReplyWith(res);
-					break;
-				default:
-					throw new ArgumentOutOfRangeException($"Unknown ReadAllResult: {res.Result}");
-			}
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token && cts.IsTimedOut)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.FilteredReadAllEventsForwardCompleted(
-					msg.CorrelationId, FilteredReadAllResult.Expired,
-					default, ResolvedEvent.EmptyArray, default, default, default,
-					currentPos: new TFPos(msg.CommitPosition, msg.PreparePosition),
-					TFPos.Invalid, TFPos.Invalid, default, default, default));
-			}
-
-			Log.Debug(
-				"Read All Stream Events Forward Filtered operation has expired for C:{0}/P:{1}. Operation Expired at {2}",
-				msg.CommitPosition, msg.PreparePosition, msg.Expires);
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token)
-		{
-			throw new OperationCanceledException(null, ex, cts.CancellationOrigin);
-		}
-	}
-
-	async ValueTask IAsyncHandle<ClientMessage.FilteredReadAllEventsBackward>.HandleAsync(
-		ClientMessage.FilteredReadAllEventsBackward msg, CancellationToken token)
-	{
-		if (msg.CancellationToken.IsCancellationRequested)
-		{
-			return;
-		}
-
-		if (msg.Expires < DateTime.UtcNow)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.FilteredReadAllEventsBackwardCompleted(
-					msg.CorrelationId, FilteredReadAllResult.Expired,
-					default, ResolvedEvent.EmptyArray, default, default, default,
-					currentPos: new TFPos(msg.CommitPosition, msg.PreparePosition),
-					TFPos.Invalid, TFPos.Invalid, default, default));
-			}
-
-			Log.Debug(
-				"Read All Stream Events Backward Filtered operation has expired for C:{0}/P:{1}. Operation Expired at {2}",
-				msg.CommitPosition, msg.PreparePosition, msg.Expires);
-			return;
-		}
-
-		using var cts = Multiplex(ref token, msg);
-		try
-		{
-			var res = await FilteredReadAllEventsBackward(msg, token);
-			switch (res.Result)
-			{
-				case FilteredReadAllResult.Success:
-					if (msg.LongPollTimeout.HasValue && res.IsEndOfStream && res.Events.Count is 0)
-					{
-						_publisher.Publish(new SubscriptionMessage.PollStream(
-							SubscriptionsService.AllStreamsSubscriptionId, res.TfLastCommitPosition, null,
-							DateTime.UtcNow + msg.LongPollTimeout.Value, msg));
-					}
-					else
-					{
-						msg.Envelope.ReplyWith(res);
-					}
-
-					break;
-				case FilteredReadAllResult.NotModified:
-					if (msg.LongPollTimeout.HasValue && res.IsEndOfStream &&
-						res.CurrentPos.CommitPosition > res.TfLastCommitPosition)
-					{
-						_publisher.Publish(new SubscriptionMessage.PollStream(
-							SubscriptionsService.AllStreamsSubscriptionId, res.TfLastCommitPosition, null,
-							DateTime.UtcNow + msg.LongPollTimeout.Value, msg));
-					}
-					else
-					{
-						msg.Envelope.ReplyWith(res);
-					}
-
-					break;
-				case FilteredReadAllResult.Error:
-				case FilteredReadAllResult.AccessDenied:
-				case FilteredReadAllResult.InvalidPosition:
-					msg.Envelope.ReplyWith(res);
-					break;
-				default:
-					throw new ArgumentOutOfRangeException($"Unknown ReadAllResult: {res.Result}");
-			}
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token && cts.IsTimedOut)
-		{
-			if (msg.ReplyOnExpired)
-			{
-				msg.Envelope.ReplyWith(new ClientMessage.FilteredReadAllEventsBackwardCompleted(
-					msg.CorrelationId, FilteredReadAllResult.Expired,
-					default, ResolvedEvent.EmptyArray, default, default, default,
-					currentPos: new TFPos(msg.CommitPosition, msg.PreparePosition),
-					TFPos.Invalid, TFPos.Invalid, default, default));
-			}
-
-			Log.Debug(
-				"Read All Stream Events Backward Filtered operation has expired for C:{0}/P:{1}. Operation Expired at {2}",
-				msg.CommitPosition, msg.PreparePosition, msg.Expires);
-		}
-		catch (OperationCanceledException ex) when (ex.CancellationToken == cts.Token)
-		{
-			throw new OperationCanceledException(null, ex, cts.CancellationOrigin);
-		}
-	}
 
 	async ValueTask IAsyncHandle<StorageMessage.EffectiveStreamAclRequest>.HandleAsync(
 		StorageMessage.EffectiveStreamAclRequest msg, CancellationToken token)
@@ -614,434 +97,6 @@ public class StorageReaderWorker<TStreamId> :
 		}
 
 		msg.Envelope.ReplyWith(reply);
-	}
-
-	private async ValueTask<ClientMessage.ReadEventCompleted> ReadEvent(ClientMessage.ReadEvent msg,
-		CancellationToken token)
-	{
-		try
-		{
-			var streamName = msg.EventStreamId;
-			var streamId = _readIndex.GetStreamId(streamName);
-			var result = await _readIndex.ReadEvent(streamName, streamId, msg.EventNumber, token);
-			var record = result.Result is ReadEventResult.Success && msg.ResolveLinkTos
-				? await ResolveLinkToEvent(result.Record, msg.User, null, token)
-				: ResolvedEvent.ForUnresolvedEvent(result.Record);
-			if (record is null)
-			{
-				return NoData(msg, ReadEventResult.AccessDenied);
-			}
-
-			if (result.Result is ReadEventResult.NoStream or ReadEventResult.NotFound &&
-				_systemStreams.IsMetaStream(streamId) &&
-				result.OriginalStreamExists.HasValue &&
-				result.OriginalStreamExists.Value)
-			{
-				return NoData(msg, ReadEventResult.Success);
-			}
-
-			return new ClientMessage.ReadEventCompleted(msg.CorrelationId, msg.EventStreamId, result.Result,
-				record.Value, result.Metadata, false, null);
-		}
-		catch (Exception exc) when (exc is not OperationCanceledException oce || oce.CancellationToken != token)
-		{
-			Log.Error(exc, "Error during processing ReadEvent request.");
-			return NoData(msg, ReadEventResult.Error, exc.Message);
-		}
-	}
-
-	private async ValueTask<ClientMessage.ReadStreamEventsForwardCompleted> ReadStreamEventsForward(
-		ClientMessage.ReadStreamEventsForward msg, CancellationToken token)
-	{
-
-		var lastIndexPosition = _readIndex.LastIndexedPosition;
-		try
-		{
-			if (msg.MaxCount > MaxPageSize)
-			{
-				throw new ArgumentException($"Read size too big, should be less than {MaxPageSize} items");
-			}
-
-			var streamName = msg.EventStreamId;
-			var streamId = _readIndex.GetStreamId(msg.EventStreamId);
-			if (msg.ValidationStreamVersion.HasValue &&
-				await _readIndex.GetStreamLastEventNumber(streamId, token) == msg.ValidationStreamVersion)
-			{
-				return NoData(msg, ReadStreamResult.NotModified, lastIndexPosition,
-					msg.ValidationStreamVersion.Value);
-			}
-
-			var result =
-				await _readIndex.ReadStreamEventsForward(streamName, streamId, msg.FromEventNumber, msg.MaxCount,
-					token);
-			CheckEventsOrder(msg, result);
-			if (await ResolveLinkToEvents(result.Records, msg.ResolveLinkTos, msg.User, token) is not { } resolvedPairs)
-			{
-				return NoData(msg, ReadStreamResult.AccessDenied, lastIndexPosition);
-			}
-
-			return new ClientMessage.ReadStreamEventsForwardCompleted(
-				msg.CorrelationId, msg.EventStreamId, msg.FromEventNumber, msg.MaxCount,
-				(ReadStreamResult)result.Result, resolvedPairs, result.Metadata, false, string.Empty,
-				result.NextEventNumber, result.LastEventNumber, result.IsEndOfStream, lastIndexPosition);
-		}
-		catch (Exception exc) when (exc is not OperationCanceledException oce || oce.CancellationToken != token)
-		{
-			Log.Error(exc, "Error during processing ReadStreamEventsForward request.");
-			return NoData(msg, ReadStreamResult.Error, lastIndexPosition, error: exc.Message);
-		}
-	}
-
-	private async ValueTask<ClientMessage.ReadStreamEventsBackwardCompleted> ReadStreamEventsBackward(
-		ClientMessage.ReadStreamEventsBackward msg, CancellationToken token)
-	{
-
-		var lastIndexedPosition = _readIndex.LastIndexedPosition;
-		try
-		{
-			if (msg.MaxCount > MaxPageSize)
-			{
-				throw new ArgumentException($"Read size too big, should be less than {MaxPageSize} items");
-			}
-
-			var streamName = msg.EventStreamId;
-			var streamId = _readIndex.GetStreamId(msg.EventStreamId);
-			if (msg.ValidationStreamVersion.HasValue &&
-				await _readIndex.GetStreamLastEventNumber(streamId, token) == msg.ValidationStreamVersion)
-			{
-				return NoData(msg, ReadStreamResult.NotModified, lastIndexedPosition,
-					msg.ValidationStreamVersion.Value);
-			}
-
-			var result = await _readIndex.ReadStreamEventsBackward(streamName, streamId, msg.FromEventNumber,
-				msg.MaxCount, token);
-			CheckEventsOrder(msg, result);
-			if (await ResolveLinkToEvents(result.Records, msg.ResolveLinkTos, msg.User, token) is not { } resolvedPairs)
-			{
-				return NoData(msg, ReadStreamResult.AccessDenied, lastIndexedPosition);
-			}
-
-			return new ClientMessage.ReadStreamEventsBackwardCompleted(
-				msg.CorrelationId, msg.EventStreamId, result.FromEventNumber, result.MaxCount,
-				(ReadStreamResult)result.Result, resolvedPairs, result.Metadata, false, string.Empty,
-				result.NextEventNumber, result.LastEventNumber, result.IsEndOfStream, lastIndexedPosition);
-		}
-		catch (Exception exc) when (exc is not OperationCanceledException oce || oce.CancellationToken != token)
-		{
-			Log.Error(exc, "Error during processing ReadStreamEventsBackward request.");
-			return NoData(msg, ReadStreamResult.Error, lastIndexedPosition, error: exc.Message);
-		}
-	}
-
-	private async ValueTask<ClientMessage.ReadAllEventsForwardCompleted>
-		ReadAllEventsForward(ClientMessage.ReadAllEventsForward msg, CancellationToken token)
-	{
-
-		var pos = new TFPos(msg.CommitPosition, msg.PreparePosition);
-		var lastIndexedPosition = _readIndex.LastIndexedPosition;
-		try
-		{
-			if (msg.MaxCount > MaxPageSize)
-			{
-				throw new ArgumentException($"Read size too big, should be less than {MaxPageSize} items");
-			}
-
-			if (pos == TFPos.HeadOfTf)
-			{
-				var checkpoint = _writerCheckpoint.Read();
-				pos = new TFPos(checkpoint, checkpoint);
-			}
-
-			if (pos.CommitPosition < 0 || pos.PreparePosition < 0)
-			{
-				return NoData(msg, ReadAllResult.InvalidPosition, pos, lastIndexedPosition, "Invalid position.");
-			}
-
-			if (msg.ValidationTfLastCommitPosition == lastIndexedPosition)
-			{
-				return NoData(msg, ReadAllResult.NotModified, pos, lastIndexedPosition);
-			}
-
-			var res = await _readIndex.ReadAllEventsForward(pos, msg.MaxCount, token);
-			if (await ResolveReadAllResult(res.Records, msg.ResolveLinkTos, msg.User, token) is not { } resolved)
-			{
-				return NoData(msg, ReadAllResult.AccessDenied, pos, lastIndexedPosition);
-			}
-
-			var metadata = await _readIndex.GetStreamMetadata(_systemStreams.AllStream, token);
-			return new ClientMessage.ReadAllEventsForwardCompleted(
-				msg.CorrelationId, ReadAllResult.Success, null, resolved, metadata, false, msg.MaxCount,
-				res.CurrentPos, res.NextPos, res.PrevPos, lastIndexedPosition);
-		}
-		catch (Exception exc) when (exc is InvalidReadException or UnableToReadPastEndOfStreamException)
-		{
-			Log.Warning(exc,
-				"Error during processing ReadAllEventsBackward request. The read appears to be at an invalid position.");
-			return NoData(msg, ReadAllResult.InvalidPosition, pos, lastIndexedPosition, exc.Message);
-		}
-		catch (Exception exc) when (exc is not OperationCanceledException oce || oce.CancellationToken != token)
-		{
-			Log.Error(exc, "Error during processing ReadAllEventsForward request.");
-			return NoData(msg, ReadAllResult.Error, pos, lastIndexedPosition, exc.Message);
-		}
-	}
-
-	private async ValueTask<ClientMessage.ReadAllEventsBackwardCompleted> ReadAllEventsBackward(
-		ClientMessage.ReadAllEventsBackward msg, CancellationToken token)
-	{
-
-		var pos = new TFPos(msg.CommitPosition, msg.PreparePosition);
-		var lastIndexedPosition = _readIndex.LastIndexedPosition;
-		try
-		{
-			if (msg.MaxCount > MaxPageSize)
-			{
-				throw new ArgumentException($"Read size too big, should be less than {MaxPageSize} items");
-			}
-
-			if (pos == TFPos.HeadOfTf)
-			{
-				var checkpoint = _writerCheckpoint.Read();
-				pos = new TFPos(checkpoint, checkpoint);
-			}
-
-			if (pos.CommitPosition < 0 || pos.PreparePosition < 0)
-			{
-				return NoData(msg, ReadAllResult.InvalidPosition, pos, lastIndexedPosition, "Invalid position.");
-			}
-
-			if (msg.ValidationTfLastCommitPosition == lastIndexedPosition)
-			{
-				return NoData(msg, ReadAllResult.NotModified, pos, lastIndexedPosition);
-			}
-
-			var res = await _readIndex.ReadAllEventsBackward(pos, msg.MaxCount, token);
-			if (await ResolveReadAllResult(res.Records, msg.ResolveLinkTos, msg.User, token) is not { } resolved)
-			{
-				return NoData(msg, ReadAllResult.AccessDenied, pos, lastIndexedPosition);
-			}
-
-			var metadata = await _readIndex.GetStreamMetadata(_systemStreams.AllStream, token);
-			return new ClientMessage.ReadAllEventsBackwardCompleted(
-				msg.CorrelationId, ReadAllResult.Success, null, resolved, metadata, false, msg.MaxCount,
-				res.CurrentPos, res.NextPos, res.PrevPos, lastIndexedPosition);
-		}
-		catch (Exception exc) when (exc is InvalidReadException or UnableToReadPastEndOfStreamException)
-		{
-			Log.Warning(exc,
-				"Error during processing ReadAllEventsBackward request. The read appears to be at an invalid position.");
-			return NoData(msg, ReadAllResult.InvalidPosition, pos, lastIndexedPosition, exc.Message);
-		}
-		catch (Exception exc) when (exc is not OperationCanceledException oce || oce.CancellationToken != token)
-		{
-			Log.Error(exc, "Error during processing ReadAllEventsBackward request.");
-			return NoData(msg, ReadAllResult.Error, pos, lastIndexedPosition, exc.Message);
-		}
-	}
-
-	private async ValueTask<ClientMessage.FilteredReadAllEventsForwardCompleted> FilteredReadAllEventsForward(
-		ClientMessage.FilteredReadAllEventsForward msg, CancellationToken token)
-	{
-
-		var pos = new TFPos(msg.CommitPosition, msg.PreparePosition);
-		var lastIndexedPosition = _readIndex.LastIndexedPosition;
-		try
-		{
-			if (msg.MaxCount > MaxPageSize)
-			{
-				throw new ArgumentException($"Read size too big, should be less than {MaxPageSize} items");
-			}
-
-			if (pos == TFPos.HeadOfTf)
-			{
-				var checkpoint = _writerCheckpoint.Read();
-				pos = new TFPos(checkpoint, checkpoint);
-			}
-
-			if (pos.CommitPosition < 0 || pos.PreparePosition < 0)
-			{
-				return NoDataForFilteredCommand(msg, FilteredReadAllResult.InvalidPosition, pos, lastIndexedPosition,
-					"Invalid position.");
-			}
-
-			if (msg.ValidationTfLastCommitPosition == lastIndexedPosition)
-			{
-				return NoDataForFilteredCommand(msg, FilteredReadAllResult.NotModified, pos,
-					lastIndexedPosition);
-			}
-
-			var res = await _readIndex.ReadAllEventsForwardFiltered(pos, msg.MaxCount, msg.MaxSearchWindow,
-				msg.EventFilter, token);
-			if (await ResolveReadAllResult(res.Records, msg.ResolveLinkTos, msg.User, token) is not { } resolved)
-			{
-				return NoDataForFilteredCommand(msg, FilteredReadAllResult.AccessDenied, pos,
-					lastIndexedPosition);
-			}
-
-			var metadata = await _readIndex.GetStreamMetadata(_systemStreams.AllStream, token);
-			return new ClientMessage.FilteredReadAllEventsForwardCompleted(
-				msg.CorrelationId, FilteredReadAllResult.Success, null, resolved, metadata, false,
-				msg.MaxCount,
-				res.CurrentPos, res.NextPos, res.PrevPos, lastIndexedPosition, res.IsEndOfStream,
-				res.ConsideredEventsCount);
-		}
-		catch (Exception exc) when (exc is InvalidReadException or UnableToReadPastEndOfStreamException)
-		{
-			Log.Warning(exc,
-				"Error during processing ReadAllEventsForwardFiltered request. The read appears to be at an invalid position.");
-			return NoDataForFilteredCommand(msg, FilteredReadAllResult.InvalidPosition, pos, lastIndexedPosition,
-				exc.Message);
-		}
-		catch (Exception exc) when (exc is not OperationCanceledException oce || oce.CancellationToken != token)
-		{
-			Log.Error(exc, "Error during processing ReadAllEventsForwardFiltered request.");
-			return NoDataForFilteredCommand(msg, FilteredReadAllResult.Error, pos, lastIndexedPosition,
-				exc.Message);
-		}
-	}
-
-	private async ValueTask<ClientMessage.FilteredReadAllEventsBackwardCompleted> FilteredReadAllEventsBackward(
-		ClientMessage.FilteredReadAllEventsBackward msg, CancellationToken token)
-	{
-
-		var pos = new TFPos(msg.CommitPosition, msg.PreparePosition);
-		var lastIndexedPosition = _readIndex.LastIndexedPosition;
-		try
-		{
-			if (msg.MaxCount > MaxPageSize)
-			{
-				throw new ArgumentException($"Read size too big, should be less than {MaxPageSize} items");
-			}
-
-			if (pos == TFPos.HeadOfTf)
-			{
-				var checkpoint = _writerCheckpoint.Read();
-				pos = new TFPos(checkpoint, checkpoint);
-			}
-
-			if (pos.CommitPosition < 0 || pos.PreparePosition < 0)
-			{
-				return NoDataForFilteredCommand(msg, FilteredReadAllResult.InvalidPosition, pos, lastIndexedPosition,
-					"Invalid position.");
-			}
-
-			if (msg.ValidationTfLastCommitPosition == lastIndexedPosition)
-			{
-				return NoDataForFilteredCommand(msg, FilteredReadAllResult.NotModified, pos,
-					lastIndexedPosition);
-			}
-
-			var res = await _readIndex.ReadAllEventsBackwardFiltered(pos, msg.MaxCount, msg.MaxSearchWindow,
-				msg.EventFilter, token);
-			if (await ResolveReadAllResult(res.Records, msg.ResolveLinkTos, msg.User, token) is not { } resolved)
-			{
-				return NoDataForFilteredCommand(msg, FilteredReadAllResult.AccessDenied, pos,
-					lastIndexedPosition);
-			}
-
-			var metadata = await _readIndex.GetStreamMetadata(_systemStreams.AllStream, token);
-			return new ClientMessage.FilteredReadAllEventsBackwardCompleted(
-				msg.CorrelationId, FilteredReadAllResult.Success, null, resolved, metadata, false,
-				msg.MaxCount,
-				res.CurrentPos, res.NextPos, res.PrevPos, lastIndexedPosition, res.IsEndOfStream);
-		}
-		catch (Exception exc) when (exc is InvalidReadException or UnableToReadPastEndOfStreamException)
-		{
-			Log.Warning(exc,
-				"Error during processing ReadAllEventsBackwardFiltered request. The read appears to be at an invalid position.");
-			return NoDataForFilteredCommand(msg, FilteredReadAllResult.InvalidPosition, pos, lastIndexedPosition,
-				exc.Message);
-		}
-		catch (Exception exc) when (exc is not OperationCanceledException oce || oce.CancellationToken != token)
-		{
-			Log.Error(exc, "Error during processing ReadAllEventsBackwardFiltered request.");
-			return NoDataForFilteredCommand(msg, FilteredReadAllResult.Error, pos, lastIndexedPosition,
-				exc.Message);
-		}
-	}
-
-	private static ClientMessage.ReadEventCompleted NoData(ClientMessage.ReadEvent msg, ReadEventResult result,
-		string error = null)
-	{
-		return new ClientMessage.ReadEventCompleted(msg.CorrelationId, msg.EventStreamId, result,
-			ResolvedEvent.EmptyEvent, null, false, error);
-	}
-
-	private static ClientMessage.ReadStreamEventsForwardCompleted NoData(ClientMessage.ReadStreamEventsForward msg,
-		ReadStreamResult result, long lastIndexedPosition, long lastEventNumber = -1, string error = null)
-	{
-		return new ClientMessage.ReadStreamEventsForwardCompleted(
-			msg.CorrelationId, msg.EventStreamId, msg.FromEventNumber, msg.MaxCount, result,
-			EmptyRecords, null, false, error ?? string.Empty, -1, lastEventNumber, true, lastIndexedPosition);
-	}
-
-	private static ClientMessage.ReadStreamEventsBackwardCompleted NoData(
-		ClientMessage.ReadStreamEventsBackward msg, ReadStreamResult result, long lastIndexedPosition,
-		long lastEventNumber = -1, string error = null)
-	{
-		return new ClientMessage.ReadStreamEventsBackwardCompleted(
-			msg.CorrelationId, msg.EventStreamId, msg.FromEventNumber, msg.MaxCount, result,
-			EmptyRecords, null, false, error ?? string.Empty, -1, lastEventNumber, true, lastIndexedPosition);
-	}
-
-	private ClientMessage.ReadAllEventsForwardCompleted NoData(ClientMessage.ReadAllEventsForward msg,
-		ReadAllResult result, TFPos pos, long lastIndexedPosition, string error = null)
-	{
-		return new ClientMessage.ReadAllEventsForwardCompleted(
-			msg.CorrelationId, result, error, ResolvedEvent.EmptyArray, null, false,
-			msg.MaxCount, pos, TFPos.Invalid, TFPos.Invalid, lastIndexedPosition);
-	}
-
-	private ClientMessage.FilteredReadAllEventsForwardCompleted NoDataForFilteredCommand(
-		ClientMessage.FilteredReadAllEventsForward msg, FilteredReadAllResult result, TFPos pos,
-		long lastIndexedPosition, string error = null)
-	{
-		return new ClientMessage.FilteredReadAllEventsForwardCompleted(
-			msg.CorrelationId, result, error, ResolvedEvent.EmptyArray, null, false,
-			msg.MaxCount, pos, TFPos.Invalid, TFPos.Invalid, lastIndexedPosition, false, 0L);
-	}
-
-	private ClientMessage.FilteredReadAllEventsBackwardCompleted NoDataForFilteredCommand(
-		ClientMessage.FilteredReadAllEventsBackward msg, FilteredReadAllResult result, TFPos pos,
-		long lastIndexedPosition, string error = null)
-	{
-		return new ClientMessage.FilteredReadAllEventsBackwardCompleted(
-			msg.CorrelationId, result, error, ResolvedEvent.EmptyArray, null, false,
-			msg.MaxCount, pos, TFPos.Invalid, TFPos.Invalid, lastIndexedPosition, false);
-	}
-
-	private ClientMessage.ReadAllEventsBackwardCompleted NoData(ClientMessage.ReadAllEventsBackward msg,
-		ReadAllResult result, TFPos pos, long lastIndexedPosition, string error = null)
-	{
-		return new ClientMessage.ReadAllEventsBackwardCompleted(
-			msg.CorrelationId, result, error, ResolvedEvent.EmptyArray, null, false,
-			msg.MaxCount, pos, TFPos.Invalid, TFPos.Invalid, lastIndexedPosition);
-	}
-
-	private static void CheckEventsOrder(ClientMessage.ReadStreamEventsForward msg, IndexReadStreamResult result)
-	{
-		for (var index = 1; index < result.Records.Length; index++)
-		{
-			if (result.Records[index].EventNumber != result.Records[index - 1].EventNumber + 1)
-			{
-				throw new Exception(
-					$"Invalid order of events has been detected in read index for the event stream '{msg.EventStreamId}'. " +
-					$"The event {result.Records[index].EventNumber} at position {result.Records[index].LogPosition} goes after the event {result.Records[index - 1].EventNumber} at position {result.Records[index - 1].LogPosition}");
-			}
-		}
-	}
-
-	private static void CheckEventsOrder(ClientMessage.ReadStreamEventsBackward msg, IndexReadStreamResult result)
-	{
-		for (var index = 1; index < result.Records.Length; index++)
-		{
-			if (result.Records[index].EventNumber != result.Records[index - 1].EventNumber - 1)
-			{
-				throw new Exception(
-					$"Invalid order of events has been detected in read index for the event stream '{msg.EventStreamId}'. " +
-					$"The event {result.Records[index].EventNumber} at position {result.Records[index].LogPosition} goes after the event {result.Records[index - 1].EventNumber} at position {result.Records[index - 1].LogPosition}");
-			}
-		}
 	}
 
 	private async ValueTask<IReadOnlyList<ResolvedEvent>> ResolveLinkToEvents(IReadOnlyList<EventRecord> records,
